pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Book_Recommender"
version = "0.0.1"
description = "A small local package for ML based books recommendations"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Shyam Mohan Tripathi", email = "rkknightx@gmail.com"}
]
keywords = ["machine-learning", "recommendation-system", "collaborative-filtering", "books"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "scikit-learn==1.3.2",
    "pandas==2.1.4",
    "numpy==1.26.2",
    "scipy==1.11.4",
    "flask==3.0.0",
    "PyYAML",
    "pydantic==2.5.0",
]

[project.urls]
Homepage = "https://github.com/smt2208/Book_Recommender"
Repository = "https://github.com/smt2208/Book_Recommender.git"
"Bug Tracker" = "https://github.com/smt2208/Book_Recommender/issues"

[tool.setuptools]
packages = ["Book_Recommender"]

[tool.setuptools.package-dir]
"Book_Recommender" = "."
